requirements.txt
# Generated by Cursor AI for Maverick Aim Rush
django
djangorestframework
djangorestframework-simplejwt
django-cors-headers
python-dotenv
django-filter
